zap: 스터디 데이터 중복 코드 제거
